+++ permutect/test/tools/test_train_base_model.py
@@ -19,6 +19,7 @@ def test_train_base_model():
     setattr(train_model_args, constants.NUM_SELF_ATTENTION_LAYERS_NAME, 2)
     setattr(train_model_args, constants.INFO_LAYERS_NAME, [10, 10])
     setattr(train_model_args, constants.AGGREGATION_LAYERS_NAME, [20, 20, 20])
+    setattr(train_model_args, constants.CALIBRATION_LAYERS_NAME, [10, 10, 10])
     cnn_layer_strings = ['convolution/kernel_size=3/out_channels=64',
                      'pool/kernel_size=2',
                      'leaky_relu',
